Move math hints to math_utils
