+++ src/vm/hints/mod.rs
@@ -6,6 +6,7 @@ pub mod execute_hint;
 pub mod find_element_hint;
 pub mod hint_utils;
 pub mod keccak_utils;
+pub mod math_utils;
 pub mod memset_utils;
 pub mod pow_utils;
 pub mod secp;
